# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Gift Multiverse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a buy order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: getUserOrders
      tags: [orders]
      summary: Get current user's orders
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── SELL REQUESTS ─────────────────────────────────────────────────────────
  /sell-requests:
    post:
      operationId: createSellRequest
      tags: [sellRequests]
      summary: Submit a sell request
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellRequestInput"
      responses:
        "201":
          description: Sell request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellRequest"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: getUserSellRequests
      tags: [sellRequests]
      summary: Get current user's sell requests
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User sell requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SellRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── DASHBOARD ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    OrderInput:
      type: object
      required: [giftCardId, quantity]
      properties:
        giftCardId:
          type: integer
        quantity:
          type: integer
          minimum: 1
    Order:
      type: object
      required: [id, userId, giftCardId, quantity, totalPrice, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        giftCardId:
          type: integer
        giftCardName:
          type: ["string", "null"]
        giftCardBrand:
          type: ["string", "null"]
        quantity:
          type: integer
        totalPrice:
          type: number
        status:
          type: string
          enum: [pending, processing, completed, cancelled]
        paymentMethod:
          type: ["string", "null"]
        paymentRef:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    SellRequestInput:
      type: object
      required: [brand, websiteUrl, cardFaceValue, askingPrice, voucherCode, cardType]
      properties:
        brand:
          type: string
        websiteUrl:
          type: string
        cardFaceValue:
          type: number
        askingPrice:
          type: number
        voucherCode:
          type: string
        cardType:
          type: string
          enum: [digital, physical]
        extraDetails:
          type: ["string", "null"]
    SellRequest:
      type: object
      required: [id, userId, brand, websiteUrl, cardFaceValue, askingPrice, voucherCode, cardType, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        brand:
          type: string
        websiteUrl:
          type: string
        cardFaceValue:
          type: number
        askingPrice:
          type: number
        voucherCode:
          type: string
        cardType:
          type: string
        extraDetails:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        approvedPayout:
          type: ["number", "null"]
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...