# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gift Multiverse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  userId: number;
  giftCardId: number;
  /** @nullable */
  giftCardName?: string | null;
  /** @nullable */
  giftCardBrand?: string | null;
  quantity: number;
  totalPrice: number;
  status: OrderStatus;
  /** @nullable */
  paymentMethod?: string | null;
  /** @nullable */
  paymentRef?: string | null;
  createdAt: Date;
}