# lib/api-zod/src/generated/types/sellRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Gift Multiverse API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SellRequestStatus } from './sellRequestStatus';

export interface SellRequest {
  id: number;
  userId: number;
  brand: string;
  websiteUrl: string;
  cardFaceValue: number;
  askingPrice: number;
  voucherCode: string;
  cardType: string;
  /** @nullable */
  extraDetails?: string | null;
  status: SellRequestStatus;
  /** @nullable */
  approvedPayout?: number | null;
  /** @nullable */
  adminNote?: string | null;
  createdAt: Date;
}